use shared rollup config to reduce duplication
