+++ packages-exp/functions-exp/rollup.config.js
@@ -18,11 +18,7 @@
 import json from 'rollup-plugin-json';
 import typescriptPlugin from 'rollup-plugin-typescript2';
 import typescript from 'typescript';
-import pkg from './package.json';
-
-const deps = Object.keys(
-  Object.assign({}, pkg.peerDependencies, pkg.dependencies)
-);
+import { es2017BuildsNoPlugin, es5BuildsNoPlugin } from './rollup.shared';
 
 /**
  * ES5 Builds
@@ -34,27 +30,14 @@ const es5BuildPlugins = [
   json()
 ];
 
-const es5Builds = [
-  /**
-   * Browser Builds
-   */
-  {
-    input: 'src/index.ts',
-    output: [{ file: pkg.module, format: 'es', sourcemap: true }],
-    plugins: es5BuildPlugins,
-    external: id => deps.some(dep => id === dep || id.startsWith(`${dep}/`))
-  },
-  /**
-   * Node.js Build
-   */
-  {
-    input: 'src/index.node.ts',
-    output: [{ file: pkg.main, format: 'cjs', sourcemap: true }],
-    plugins: es5BuildPlugins,
-    external: id => deps.some(dep => id === dep || id.startsWith(`${dep}/`))
-  }
-];
+const es5Builds = es5BuildsNoPlugin.map(build => ({
+  ...build,
+  plugins: es5BuildPlugins
+}));
 
+/**
+ * ES2017 Builds
+ */
 const es2017BuildPlugins = [
   typescriptPlugin({
     typescript,
@@ -67,23 +50,9 @@ const es2017BuildPlugins = [
   json({ preferConst: true })
 ];
 
-/**
- * ES2017 Builds
- */
-const es2017Builds = [
-  {
-    /**
-     * Browser Build
-     */
-    input: 'src/index.ts',
-    output: {
-      file: pkg.esm2017,
-      format: 'es',
-      sourcemap: true
-    },
-    plugins: es2017BuildPlugins,
-    external: id => deps.some(dep => id === dep || id.startsWith(`${dep}/`))
-  }
-];
+const es2017Builds = es2017BuildsNoPlugin.map(build => ({
+  ...build,
+  plugins: es2017BuildPlugins
+}));
 
 export default [...es5Builds, ...es2017Builds];
