+++ packages-exp/installations-exp/package.json
@@ -15,6 +15,7 @@
     "lint:fix": "eslint --fix -c .eslintrc.js '**/*.ts' --ignore-path '../../.gitignore'",
     "build": "rollup -c && yarn api-report",
     "build:deps": "lerna run --scope @firebase/installations-exp --include-dependencies build",
+    "build:release": "rollup -c rollup.config.release.js && yarn api-report",
     "dev": "rollup -c -w",
     "test": "yarn type-check && yarn test:karma && yarn lint",
     "test:ci": "node ../../scripts/run_tests_in_ci.js",
@@ -24,7 +25,7 @@
     "serve": "yarn serve:build && yarn serve:host",
     "serve:build": "rollup -c test-app/rollup.config.js",
     "serve:host": "http-server -c-1 test-app",
-    "prepare": "yarn build",
+    "prepare": "yarn build:release",
     "api-report": "api-extractor run --local --verbose",
     "predoc": "node ../../scripts/exp/remove-exp.js temp",
     "doc": "api-documenter markdown --input temp --output docs",
