+++ packages-exp/performance-exp/package.json
@@ -14,12 +14,13 @@
     "lint:fix": "eslint --fix -c .eslintrc.js '**/*.ts'",
     "build": "rollup -c",
     "build:deps": "lerna run --scope @firebase/performance-exp --include-dependencies build",
+    "build:release": "rollup -c rollup.config.release.js",
     "dev": "rollup -c -w",
     "test": "run-p lint test:browser",
     "test:ci": "node ../../scripts/run_tests_in_ci.js -s test:browser",
     "test:browser": "karma start --single-run",
     "test:debug": "karma start --browsers=Chrome --auto-watch",
-    "prepare": "yarn build",
+    "prepare": "yarn build:release",
     "prettier": "prettier --write '{src,test}/**/*.{js,ts}'"
   },
   "peerDependencies": {
